# swagger.yaml
openapi: 3.0.3
info:
  title: ReadersHub API
  version: 1.0.0
  description: |
    API documentation for the ReadersHub backend.
    Complete CRUD operations for Books, Users, Authors, Movies, and Reviews.
servers:
  - url: http://localhost:8000
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ReviewCreate:
      type: object
      required:
        - review
        - rating
      properties:
        review:
          type: string
        rating:
          type: number
          format: float
...
paths:
...
  /api/v1/books/{bookId}/reviews:
    get:
      tags: [Reviews]
      summary: Get all reviews for a book
      security:
        - bearerAuth: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of reviews
    post:
      tags: [Reviews]
      summary: Create review for a book (user only)
      security:
        - bearerAuth: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewCreate"
      responses:
        "201":
          description: Review created
...
  /api/v1/reviews:
    get:
      tags: [Reviews]
      summary: Get all reviews
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of all reviews
...